% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Problem}
\usage{
birthday(n)
}
\arguments{
\item{n}{Integer of class size.}
}
\value{
A numeric vector of probabilities.
}
\description{
Computes the probability of two people in a group of size n
sharing the same birthday.
}
\examples{
birthday(23)
birthday(1:20)
}
