% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data}
\format{
## `who`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
ddt
}
\description{
A subset of data over fish and rivers ...
}
\keyword{datasets}
