% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Dataset}
\format{
A data frame with 16 rows and 2 variables:
\describe{
  \item{DAMAGE}{Amount of property damage}
  \item{DISTANCE}{How far the fire spread}
}
}
\source{
FIREDAM.csv
}
\usage{
fire
}
\description{
A dataset containing fire damage records.
}
\keyword{datasets}
