% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Experiment}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Integer. Number of iterations (Simulated experiments).}

\item{n}{Integer. Number of trials in each experiment.}

\item{p}{Numerical. Probability of success in each trail.}
}
\value{
A named table of empirical proportions for 0 to n successes.
}
\description{
This function simulates binomial experiments by sampling n Bernoulli
trials with probability p of success, repeated "iter" times.
It also produces a barplot of the emperical proportions and returns
the table.
}
\examples{
mybin(iter = 100, n = 10, p = 0.5)
mybin(iter = 1000, n = 20, p = 0.3)
}
