% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My bootstrap function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap re samples (default 10000).}

\item{x}{Numeric data vector to be bootstrapped.}

\item{fun}{Statistic to compute on each bootstrap sample.}

\item{alpha}{Significance level for the interval, (0.05 for 95 CI).}

\item{cx}{Text size for plot labels.}

\item{...}{Extra arguments passed to hist().}
}
\value{
A list with components: ci (bootstrap confidence interval), fun (the
statistic used), x (the original data), and xstat (the bootstrap sample
statistics).
}
\description{
My bootstrap function
}
\examples{
\dontrun{
sam <- rnorm(20, mean = 10, sd = 4)
myboot2(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5)
}
}
