% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{My Central Limit Function
Notice that I have assigned default values which can be
changed when the function is called.
Uniform.}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Sample size.}

\item{iter}{Number of iterations.}

\item{a}{Upper bound of uniform.}

\item{b}{Lower bound of uniform.}
}
\value{
Graph.
}
\description{
My Central Limit Function
Notice that I have assigned default values which can be
changed when the function is called.
Uniform.
}
\examples{
mycltu(n=20,iter=100000)
}
