% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p, make_plots = TRUE)
}
\arguments{
\item{N}{integer, number of seats.}

\item{gamma}{probability tolerance for true overbooking.}

\item{p}{actual attendees probability.}

\item{make_plots}{plot discrete and continuous objectives.}
}
\value{
named list containing nd, nc, N, p and gamma - where nd is
calculated using the discrete distribution and nc is the same calculated with
normal approximation.
}
\description{
Package for overbooking problem.
}
\examples{
ntickets(400, 0.02, 0.95)
}
