% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_values.R
\name{z_values}
\alias{z_values}
\title{Standardize a numeric vector}
\usage{
z_values(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of z values
}
\description{
This function takes a numeric vector and returns the standardized
values
}
\examples{
mpg <- c(21, 22, 19, 30, 15)
z_values(mpg)
}
